Use a 4-way parallel BLAKE2s instead of SHA-384
This is a schema-incompatible change that can't possibly be automated.
The recommended way to handle this is to make a new database, make snapshot(s),
then make snapshot(s) to the *old* database using the *old* version. If there
are no mismatches, start using the new database only.

BLAKE2s is, like SHA-384, a cryptographic hash function. Its advantages
are flexibility and speed. I measured the very implementation I'm committing
on various machines, and on every one BLAKE2sp4 was faster than SHA-384.
The following table shows the hash throughput for 16 KB input blocks on
single core:
+----------------+---------------+---------------+
|CPU @ freq.(GHz)|    SHA-384    |   BLAKE2sp4   |
+----------------+---------------+---------------+
|   Zen @ 3.5    |   630 MB/s    |   1120 MB/s   |
+----------------+---------------+---------------+
| Haswell @ 3.6  |   660 MB/s    |   1030 MB/s   |
+----------------+---------------+---------------+
|IvyBridge @ 3.3 |   440 MB/s    |   850 MB/s    |
+----------------+---------------+---------------+
| Core 2 @ 2.66  |   260 MB/s    |   520 MB/s    |
+----------------+---------------+---------------+
|Pentium M @ 1.6 |    60 MB/s    |    90 MB/s    |
+----------------+---------------+---------------+
|Pentium 3 @ 1.13|    20 MB/s    |    45 MB/s    |
+----------------+---------------+---------------+

Additionally, this lets us get rid of the OpenSSL dependency, and thus also of
its licensing exception.