Speed up garbage collection
The use of a compound SELECT in an IN operator makes SQLite unable to use the
indices on {snapshot_files,file_chunks}(hash_id) directly. Instead, it's forced
to do a full scan on those indices, inserting each ID into an ephemeral index,
then using that ephemeral index to filter out referenced IDs. This is still
O(nlogn), but it's possible to do better.

If the NOT IN (… UNION ALL …) is transformed to NOT IN (…) AND NOT IN (…), which
is a valid de Morgan's transformation, SQLite is able to use the indices
directly for both IN operators, with no ephemeral indices involved.

In practice, it results in a 25% speedup in a relatively small, just under 200
MB, database containing just 10 snapshots. The benefit is even bigger for larger
databases.