Mark statement column getter for integers as noexcept
It can indeed never throw, because the column index is checked ahead of time
when creating a cursor, and it's private so outside world can't just pass
invalid indices anyway.

The getters for strings and blobs aren't noexcept because they can fail due to
OOM (which is currently unchecked, but that may change).